% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmi510-template.R
\name{adj_R2}
\alias{adj_R2}
\title{Calculate the adjusted R-squared between predicted and ground truth continuous variables}
\usage{
adj_R2(pred, truth, n_p)
}
\arguments{
\item{pred}{A numeric vector of predicted values}

\item{truth}{A numeric vector of true values}

\item{n_p}{An integer value indicating the number of model parameters, excluding the intercept.}
}
\value{
A numeric value representing the adjusted R-squared statistic
}
\description{
This function calculates the adjusted R-squared statistic between predicted and ground truth
continuous variables, given the number of model parameters (excluding the intercept).
}
\examples{
set.seed(123)
n <- 100
x <- rnorm(n)
y <- 2 * x + rnorm(n)
model <- lm(y ~ x)
adj_R2(predict(model), y, n_p = 1)
}
