% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmi510-template.R
\name{classes}
\alias{classes}
\title{Get the classes of variables in a tibble}
\usage{
classes(x)
}
\arguments{
\item{x}{A tibble}
}
\value{
A character vector containing the classes of each variable
}
\description{
This function takes a tibble \code{x} and returns a character vector
containing the classes of each variable in the tibble.
}
\examples{
library(tibble)
x <- tibble(a = 1:5, b = c("foo", "bar", "baz", "qux", "quux"), c = TRUE)
classes(x)
}
