% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmi510-template.R
\name{df_scale}
\alias{df_scale}
\title{Scale the numeric variables in a tibble}
\usage{
df_scale(x, center = TRUE, scale = TRUE)
}
\arguments{
\item{x}{A tibble}

\item{center}{A logical indicating whether to center the data}

\item{scale}{A logical indicating whether to scale the data}
}
\value{
A tibble with numeric variables scaled
}
\description{
This function takes a tibble \code{x} and returns a tibble in which the
numeric variables have been scaled using the function \code{scale}.
}
\examples{
library(tibble)
x <- tibble(a = 1:5, b = c("foo", "bar", "baz", "qux", "quux"), c = TRUE)
df_scale(x)
}
