% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmi510-template.R
\name{f1}
\alias{f1}
\title{Calculate F1 Score}
\usage{
f1(pred, truth)
}
\arguments{
\item{pred}{A vector of predicted labels.}

\item{truth}{A vector of true labels.}
}
\value{
The F1 score.
}
\description{
This function calculates the F1 score metric based on comparing predicted and truth labels.
}
