% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmi510-template.R
\name{is_max}
\alias{is_max}
\title{Check if values in a vector are equal to the maximum value}
\usage{
is_max(x, na.rm = T)
}
\arguments{
\item{x}{A numeric vector}

\item{na.rm}{Logical; should missing values be removed from the vector? (default TRUE)}
}
\value{
A logical vector indicating whether values in x are equal to its maximum value
}
\description{
The function checks if the values in a vector are equal to its maximum value.
By default, missing values are removed from the vector.
}
\examples{
# check if values are equal to the maximum value
x <- c(1, 2, 3, 2, 1)
is_max(x) # returns FALSE FALSE TRUE FALSE FALSE

# check if values are equal to the maximum value with missing values
x_with_na <- c(1, 2, 3, NA, 2, 1)
is_max(x_with_na, na.rm = TRUE) # returns FALSE FALSE TRUE FALSE FALSE FALSE
is_max(x_with_na, na.rm = FALSE) # returns FALSE FALSE TRUE NA FALSE FALSE

}
