% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmi510-template.R
\name{is_min}
\alias{is_min}
\title{Check if values in a vector are equal to the minimum value}
\usage{
is_min(x, na.rm = T)
}
\arguments{
\item{x}{A numeric vector}

\item{na.rm}{Logical; should missing values be removed from the vector? (default TRUE)}
}
\value{
A logical vector indicating whether values in x are equal to its minimum value
}
\description{
The function checks if the values in a vector are equal to its minimum value.
By default, missing values are removed from the vector.
}
\examples{
# check if values are equal to the minimum value
x <- c(1, 2, 3, 2, 1)
is_min(x) # returns TRUE FALSE FALSE FALSE TRUE

# check if values are equal to the minimum value with missing values
x_with_na <- c(1, 2, 3, NA, 2, 1)
is_min(x_with_na, na.rm = TRUE) # returns TRUE FALSE FALSE FALSE TRUE FALSE
is_min(x_with_na, na.rm = FALSE) # returns TRUE FALSE FALSE NA FALSE TRUE

}
