% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmi510-template.R
\name{log_likelihood_chisq}
\alias{log_likelihood_chisq}
\title{Compute log-likelihood of a sample under the chi-squared density}
\usage{
log_likelihood_chisq(x, df)
}
\arguments{
\item{x}{a numeric vector of non-negative values}

\item{df}{a positive integer indicating the degrees of freedom}
}
\value{
the log-likelihood of the sample under the chi-squared distribution
}
\description{
Given a sample of non-negative values x and a positive integer degrees of freedom df,
this function computes the log-likelihood of the sample under the chi-squared distribution
with df degrees of freedom.
}
\examples{
x <- c(1, 2, 3)
log_likelihood_chisq(x, 2)
# expected output: -4.10517
log_likelihood_chisq(x, -1)
# expected output: Error: df must be a positive integer
log_likelihood_chisq(x, 2.5)
# expected output: Error: df must be a positive integer
log_likelihood_chisq(c(-1, 2, 3), 2)
# expected output: Error: x must be a numeric vector with non-negative values
}
