% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmi510-template.R
\name{log_likelihood_f}
\alias{log_likelihood_f}
\title{Compute log-likelihood of sample x under the f distribution.}
\usage{
log_likelihood_f(x, df1, df2)
}
\arguments{
\item{x}{A numeric vector of observations.}

\item{df1}{Degrees of freedom for the numerator.}

\item{df2}{Degrees of freedom for the denominator.}
}
\value{
The log-likelihood of the sample x under the F distribution with parameters df1 and df2.
}
\description{
This function computes the log-likelihood of a sample x under the F distribution with parameters df1 and df2.
}
\examples{
x <- rf(100, df1 = 5, df2 = 10)
log_likelihood_f(x, df1 = 5, df2 = 10)

}
\seealso{
\code{\link{df}}, \code{\link{logLik}}
}
