% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmi510-template.R
\name{log_likelihood_norm}
\alias{log_likelihood_norm}
\title{Log-likelihood of a normal distribution}
\usage{
log_likelihood_norm(x, mean, sd)
}
\arguments{
\item{x}{A numeric vector of observations}

\item{mean}{The mean of the normal distribution}

\item{sd}{The standard deviation of the normal distribution}
}
\value{
The log-likelihood of the sample \code{x} under a normal distribution
}
\description{
This function calculates the log-likelihood of a sample \code{x} under a normal
distribution with specified mean \code{mean} and standard deviation \code{sd}.
}
\examples{
log_likelihood_norm(c(1,2,3), mean = 2, sd = 1)
}
