% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmi510-template.R
\name{log_likelihood_t}
\alias{log_likelihood_t}
\title{Calculate Log-Likelihood of t Distribution}
\usage{
log_likelihood_t(x, df)
}
\arguments{
\item{x}{A numeric vector of observations.}

\item{df}{Degrees of freedom of the t distribution.}
}
\value{
The log-likelihood of x under the t distribution.
}
\description{
This function calculates the log-likelihood of a sample x under the t distribution with degrees of freedom df.
}
\examples{
log_likelihood_t(x = c(1.2, 0.9, -0.3), df = 4)
}
