% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmi510-template.R
\name{log_likelihood_unif}
\alias{log_likelihood_unif}
\title{Log-likelihood of a uniform distribution}
\usage{
log_likelihood_unif(x, min, max)
}
\arguments{
\item{x}{A numeric vector of observations}

\item{min}{The minimum value of the uniform distribution}

\item{max}{The maximum value of the uniform distribution}
}
\value{
The log-likelihood of the sample \code{x} under a uniform distribution
}
\description{
This function calculates the log-likelihood of a sample \code{x} under a uniform
distribution with specified minimum \code{min} and maximum \code{max}.
}
\examples{
log_likelihood_unif(c(0.5, 0.8, 1), min = 0, max = 1)
}
