% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmi510-template.R
\name{minimum_n_per_group}
\alias{minimum_n_per_group}
\title{Calculate the minimum sample size per group for a two-sample t-test}
\usage{
minimum_n_per_group(d, power = 0.8)
}
\arguments{
\item{d}{The expected Cohen's d.}

\item{power}{The desired power of the t-test (default is 0.8).}
}
\value{
The minimum sample size per group.
}
\description{
This function calculates the minimum number n per group needed for a two-sample t-test
with a given expected Cohen's d and power.
}
\examples{
minimum_n_per_group(0.5)
minimum_n_per_group(0.8, 0.9)
}
