% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmi510-template.R
\name{r2}
\alias{r2}
\title{Calculate R-squared}
\usage{
r2(pred, truth)
}
\arguments{
\item{pred}{A numeric vector of predicted values.}

\item{truth}{A numeric vector of ground truth values.}
}
\value{
A numeric value representing the R-squared statistic.
}
\description{
This function calculates the R-squared statistic between predicted and ground truth continuous variables.
}
\examples{
r2(c(1, 2, 3), c(1.1, 1.8, 2.9))
}
