% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmi510-template.R
\name{rando}
\alias{rando}
\title{Sample from an atomic vector or dataframe-like object}
\usage{
rando(x, n = 1, replace = T)
}
\arguments{
\item{x}{An atomic vector or a dataframe-like object}

\item{n}{The number of samples or rows to return (default 1)}

\item{replace}{Logical; should sampling be done with replacement? (default TRUE)}
}
\value{
An atomic vector or dataframe with n samples or rows
}
\description{
The function samples from either an atomic vector or a dataframe-like object
and returns either n samples or n rows as appropriate.
}
\examples{
# sample from an atomic vector
rando(1:10, 3, replace = TRUE)

# sample from a dataframe
df <- data.frame(x = 1:5, y = 6:10)
rando(df, 2, replace = TRUE)

# error case
rando(list(1, 2, 3), 2)

}
