% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmi510-template.R
\name{recall}
\alias{recall}
\title{Calculate Recall}
\usage{
recall(pred, truth)
}
\arguments{
\item{pred}{A vector of predicted labels}

\item{truth}{A vector of true labels}
}
\value{
A scalar value representing the recall score.
}
\description{
This function calculates the recall metric based on comparing predicted and truth labels.
Recall, also known as sensitivity or true positive rate, measures the proportion of actual positives
that are correctly identified by the model.
}
\examples{
pred <- c(1, 1, 0, 1, 0, 1)
truth <- c(1, 0, 0, 1, 1, 1)
recall(pred, truth)
#[1] 0.75

}
