% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmi510-template.R
\name{rep_mat}
\alias{rep_mat}
\title{Replicate rows or columns of a matrix or data frame}
\usage{
rep_mat(x, M = 1, N = 1)
}
\arguments{
\item{x}{A matrix or data frame}

\item{M}{An integer specifying the number of times to replicate the rows of \code{x}}

\item{N}{An integer specifying the number of times to replicate the columns of \code{x}}
}
\value{
A matrix or data frame with replicated rows and/or columns
}
\description{
This function takes a matrix or data frame \code{x} and replicates its rows or
columns \code{M} and \code{N} times, respectively. The resulting matrix has
\code{nrow(x)*M} rows and \code{ncol(x)*N} columns.
}
\examples{
x <- matrix(1:4, nrow = 2)
rep_mat(x, M = 3) # Replicate rows 3 times
rep_mat(x, N = 2) # Replicate columns 2 times
rep_mat(x, M = 3, N = 2) # Replicate rows 3 times and columns 2 times
}
