% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmi510-template.R
\name{sensitivity}
\alias{sensitivity}
\title{Calculate sensitivity metric based on predicted and true labels}
\usage{
sensitivity(pred, truth)
}
\arguments{
\item{pred}{A numeric vector of predicted binary labels (0 or 1)}

\item{truth}{A numeric vector of true binary labels (0 or 1)}
}
\value{
A numeric value representing the sensitivity (true positive rate) of the predictions
with respect to the true labels
}
\description{
This function takes in two vectors of binary labels, the predicted labels and the true labels,
and calculates the sensitivity (true positive rate) of the predictions with respect to the true
labels. Sensitivity is the proportion of true positives (TP) out of all actual positives (TP + FN).
}
\examples{
sensitivity(c(1,1,0,1,0), c(1,0,0,1,1))
# expected output: 0.75

}
