% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmi510-template.R
\name{specificity}
\alias{specificity}
\title{Calculate specificity metric}
\usage{
specificity(pred, truth)
}
\arguments{
\item{pred}{a vector of predicted labels}

\item{truth}{a vector of true labels}
}
\value{
a numeric value representing the specificity metric
}
\description{
This function calculates the specificity metric based on comparing predicted and truth labels.
}
\examples{
specificity(c(0,0,1,1), c(1,0,1,0)) # returns 0.5
specificity(c(0,0,1,1), c(0,0,0,0)) # returns 1
specificity(c(0,0,1,1), c(1,1,1,1)) # returns 0

}
